Fix default MTU setting
line attempts to set DESIRED_MTU to the value of the fifth argument $5 of the script. However, if $5 is not provided, it defaults to 1500.

The problem arises because the same variable $5 is being used for two different purposes: MAX_CURL_ATTEMPT and DESIRED_MTU. Therefore, when you don't pass a value for MAX_CURL_ATTEMPT, it defaults DESIRED_MTU to 1500.